quiet some compiler warnings

+++ Core/Grid/Level.cc
@@ -266,16 +266,16 @@ void Level::findNodeIndexRange(IntVector& lowIndex,IntVector& highIndex) const
   Vector l=(d_spatial_range.min()-d_anchor)/d_dcell;
   Vector h=(d_spatial_range.max()-d_anchor)/d_dcell+Vector(1,1,1);
 
-  lowIndex=IntVector(l.x(),l.y(),l.z());
-  highIndex=IntVector(h.x(),h.y(),h.z());
+  lowIndex=IntVector((int)l.x(),(int)l.y(),(int)l.z());
+  highIndex=IntVector((int)h.x(),(int)h.y(),(int)h.z());
 }
 void Level::findCellIndexRange(IntVector& lowIndex,IntVector& highIndex) const
 {
   Vector l=(d_spatial_range.min()-d_anchor)/d_dcell;
   Vector h=(d_spatial_range.max()-d_anchor)/d_dcell;
 
-  lowIndex=IntVector(l.x(),l.y(),l.z());
-  highIndex=IntVector(h.x(),h.y(),h.z());
+  lowIndex=IntVector((int)l.x(),(int)l.y(),(int)l.z());
+  highIndex=IntVector((int)h.x(),(int)h.y(),(int)h.z());
 }
 
 void Level::findInteriorCellIndexRange(IntVector& lowIndex,IntVector& highIndex) const
@@ -283,8 +283,8 @@ void Level::findInteriorCellIndexRange(IntVector& lowIndex,IntVector& highIndex)
   Vector l=(d_int_spatial_range.min()-d_anchor)/d_dcell;
   Vector h=(d_int_spatial_range.max()-d_anchor)/d_dcell;
 
-  lowIndex=IntVector(l.x(),l.y(),l.z());
-  highIndex=IntVector(h.x(),h.y(),h.z());
+  lowIndex=IntVector((int)l.x(),(int)l.y(),(int)l.z());
+  highIndex=IntVector((int)h.x(),(int)h.y(),(int)h.z());
 }
 
 void Level::findInteriorNodeIndexRange(IntVector& lowIndex,IntVector& highIndex) const
@@ -292,8 +292,8 @@ void Level::findInteriorNodeIndexRange(IntVector& lowIndex,IntVector& highIndex)
   Vector l=(d_int_spatial_range.min()-d_anchor)/d_dcell;
   Vector h=(d_int_spatial_range.max()-d_anchor)/d_dcell+Vector(1,1,1);
 
-  lowIndex=IntVector(l.x(),l.y(),l.z());
-  highIndex=IntVector(h.x(),h.y(),h.z());
+  lowIndex=IntVector((int)l.x(),(int)l.y(),(int)l.z());
+  highIndex=IntVector((int)h.x(),(int)h.y(),(int)h.z());
 }
 
 long Level::totalCells() const
